requirements.txt
# PD2 MLOps — Instituto Infnet
# Python >= 3.11

# Core ML
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.12.0
joblib>=1.3.0

# MLflow (rastreamento de experimentos)
mlflow>=2.10.0

# Streamlit (interface de inferência)
streamlit>=1.32.0
plotly>=5.20.0

# Configuração
pyyaml>=6.0.0

# Jupyter (exploração e visualização — opcional)
jupyter>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
